Move serve ha test to serve folder
