+++ python/ray/tests/test_gcs_ha_e2e.py
@@ -1,11 +1,8 @@
 import pytest
 import sys
 import threading
-from time import sleep, time
+from time import sleep
 from ray._private.test_utils import wait_for_condition
-from pytest_docker_tools import container, fetch, network, volume
-from pytest_docker_tools import wrappers
-from http.client import HTTPConnection
 
 
 # If you need to debug these tests, comment in the volume
@@ -29,112 +26,6 @@ from http.client import HTTPConnection
 # Now you can re-run the test and the logs will show
 # up in /tmp/ray in the docker-in-docker container.
 # Good luck!
-
-
-class Container(wrappers.Container):
-    def ready(self):
-        self._container.reload()
-        if self.status == "exited":
-            from pytest_docker_tools.exceptions import ContainerFailed
-
-            raise ContainerFailed(
-                self,
-                f"Container {self.name} has already exited before "
-                "we noticed it was ready",
-            )
-
-        if self.status != "running":
-            return False
-
-        networks = self._container.attrs["NetworkSettings"]["Networks"]
-        for (_, n) in networks.items():
-            if not n["IPAddress"]:
-                return False
-
-        if "Ray runtime started" in super().logs():
-            return True
-        return False
-
-    def client(self):
-        port = self.ports["8000/tcp"][0]
-        return HTTPConnection(f"localhost:{port}")
-
-    def print_logs(self):
-        for (name, content) in self.get_files("/tmp"):
-            print(f"===== log start:  {name} ====")
-            print(content.decode())
-
-
-gcs_network = network(driver="bridge")
-
-redis_image = fetch(repository="redis:latest")
-
-redis = container(
-    image="{redis_image.id}",
-    network="{gcs_network.name}",
-    command=("redis-server --save 60 1 --loglevel" " warning"),
-)
-
-head_node_vol = volume()
-worker_node_vol = volume()
-head_node_container_name = "gcs" + str(int(time()))
-head_node = container(
-    image="ray_ci:v1",
-    name=head_node_container_name,
-    network="{gcs_network.name}",
-    command=[
-        "ray",
-        "start",
-        "--head",
-        "--block",
-        "--num-cpus",
-        "0",
-        # Fix the port of raylet to make sure raylet restarts at the same
-        # ip:port is treated as a different raylet.
-        "--node-manager-port",
-        "9379",
-    ],
-    volumes={"{head_node_vol.name}": {"bind": "/tmp", "mode": "rw"}},
-    environment={"RAY_REDIS_ADDRESS": "{redis.ips.primary}:6379"},
-    wrapper_class=Container,
-    ports={
-        "8000/tcp": None,
-    },
-    # volumes={
-    #     "/tmp/ray/": {"bind": "/tmp/ray/", "mode": "rw"}
-    # },
-)
-
-worker_node = container(
-    image="ray_ci:v1",
-    network="{gcs_network.name}",
-    command=[
-        "ray",
-        "start",
-        "--address",
-        f"{head_node_container_name}:6379",
-        "--block",
-        # Fix the port of raylet to make sure raylet restarts at the same
-        # ip:port is treated as a different raylet.
-        "--node-manager-port",
-        "9379",
-    ],
-    volumes={"{worker_node_vol.name}": {"bind": "/tmp", "mode": "rw"}},
-    environment={"RAY_REDIS_ADDRESS": "{redis.ips.primary}:6379"},
-    wrapper_class=Container,
-    ports={
-        "8000/tcp": None,
-    },
-    # volumes={
-    #     "/tmp/ray/": {"bind": "/tmp/ray/", "mode": "rw"}
-    # },
-)
-
-
-@pytest.fixture
-def docker_cluster(head_node, worker_node):
-    yield (head_node, worker_node)
-
 
 scripts = """
 import ray
